Add tests for new template operation
