+++ robot-core/src/test/java/org/obolibrary/robot/QuotedEntityCheckerTest.java
@@ -2,7 +2,9 @@ package org.obolibrary.robot;
 
 import static junit.framework.TestCase.assertEquals;
 
+import org.junit.Assert;
 import org.junit.Test;
+import org.semanticweb.owlapi.manchestersyntax.parser.ManchesterOWLSyntaxClassExpressionParser;
 import org.semanticweb.owlapi.model.*;
 import org.semanticweb.owlapi.util.SimpleShortFormProvider;
 
@@ -12,6 +14,14 @@ import org.semanticweb.owlapi.util.SimpleShortFormProvider;
  * @author <a href="mailto:rctauber@gmail.com">Bekcy Tauber</a>
  */
 public class QuotedEntityCheckerTest extends CoreTest {
+
+  /** Test wrapping and escaping names. */
+  @Test
+  public void testEscaping() {
+    Assert.assertEquals("testone", QuotedEntityChecker.wrap("testone"));
+    Assert.assertEquals("'test one'", QuotedEntityChecker.wrap("test one"));
+    Assert.assertEquals("5-bromo-2\\'-deoxyuridine", QuotedEntityChecker.wrap("5-bromo-2'-deoxyuridine"));
+  }
 
   /**
    * Test resolving a label from an import.
@@ -35,5 +45,45 @@ public class QuotedEntityCheckerTest extends CoreTest {
     } else {
       throw new Exception("Class 'test one' does not exist.");
     }
+  }
+
+  /**
+   * Test the QuotedEntityChecker.
+   *
+   * @throws Exception if entities cannot be found
+   */
+  @Test
+  public void testChecker() throws Exception {
+    OWLOntology simpleParts = loadOntology("/simple_parts.owl");
+    QuotedEntityChecker checker = new QuotedEntityChecker();
+    checker.addProperty(dataFactory.getRDFSLabel());
+    checker.addAll(simpleParts);
+
+    IRI iri = IRI.create(base + "simple.owl#test1");
+    OWLClass cls = dataFactory.getOWLClass(iri);
+    Assert.assertEquals(cls, checker.getOWLClass("test one"));
+    Assert.assertEquals(cls, checker.getOWLClass("'test one'"));
+    Assert.assertEquals(cls, checker.getOWLClass("Test 1"));
+    Assert.assertEquals(iri, checker.getIRI("test one", false));
+
+    IOHelper ioHelper = new IOHelper();
+    iri = ioHelper.createIRI("GO:XXXX");
+    cls = dataFactory.getOWLClass(iri);
+    checker.setIOHelper(ioHelper);
+    Assert.assertEquals(cls, checker.getOWLClass("GO:XXXX"));
+
+    System.out.println("PARSER");
+    ManchesterOWLSyntaxClassExpressionParser parser =
+      new ManchesterOWLSyntaxClassExpressionParser(
+        dataFactory, checker
+        // new org.semanticweb.owlapi.expression.ShortFormEntityChecker(
+        //    new org.semanticweb.owlapi.util.
+        //        BidirectionalShortFormProviderAdapter(
+        //            ioHelper.getPrefixManager()))
+      );
+    // assertEquals(cls, parser.parse("'test one'"));
+    Assert.assertEquals(cls, parser.parse("GO:XXXX"));
+    // checker.add(cls, "%");
+    // assertEquals("", parser.parse("%"));
   }
 }
